+++ canon-docs/src/snippets/stories-snippets.tsx
@@ -22,6 +22,9 @@ import * as RadioGroupStories from '../../../packages/canon/src/components/Radio
 import * as TabsStories from '../../../packages/canon/src/components/Tabs/Tabs.stories';
 import * as SwitchStories from '../../../packages/canon/src/components/Switch/Switch.stories';
 import * as SearchFieldStories from '../../../packages/canon/src/components/SearchField/SearchField.stories';
+import * as TooltipStories from '../../../packages/canon/src/components/Tooltip/Tooltip.stories';
+import * as SkeletonStories from '../../../packages/canon/src/components/Skeleton/Skeleton.stories';
+import * as CardStories from '../../../packages/canon/src/components/Card/Card.stories';
 
 export const BoxSnippet = ({ story }: { story: keyof typeof BoxStories }) => {
   const stories = composeStories(BoxStories);
@@ -217,6 +220,35 @@ export const SearchFieldSnippet = ({
   story: keyof typeof SearchFieldStories;
 }) => {
   const stories = composeStories(SearchFieldStories);
+  const StoryComponent = stories[story as keyof typeof stories];
+
+  return StoryComponent ? <StoryComponent /> : null;
+};
+
+export const TooltipSnippet = ({
+  story,
+}: {
+  story: keyof typeof TooltipStories;
+}) => {
+  const stories = composeStories(TooltipStories);
+  const StoryComponent = stories[story as keyof typeof stories];
+
+  return StoryComponent ? <StoryComponent /> : null;
+};
+
+export const SkeletonSnippet = ({
+  story,
+}: {
+  story: keyof typeof SkeletonStories;
+}) => {
+  const stories = composeStories(SkeletonStories);
+  const StoryComponent = stories[story as keyof typeof stories];
+
+  return StoryComponent ? <StoryComponent /> : null;
+};
+
+export const CardSnippet = ({ story }: { story: keyof typeof CardStories }) => {
+  const stories = composeStories(CardStories);
   const StoryComponent = stories[story as keyof typeof stories];
 
   return StoryComponent ? <StoryComponent /> : null;