+++ packages/canon/src/components/ButtonIcon/types.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { Breakpoint } from '@backstage/canon';
+import { Breakpoint } from '../../';
 import { ReactElement } from 'react';
 import { ButtonProps as RAButtonProps } from 'react-aria-components';
 
@@ -28,6 +28,7 @@ export interface ButtonIconProps extends RAButtonProps {
   variant?:
     | 'primary'
     | 'secondary'
-    | Partial<Record<Breakpoint, 'primary' | 'secondary'>>;
+    | 'tertiary'
+    | Partial<Record<Breakpoint, 'primary' | 'secondary' | 'tertiary'>>;
   icon?: ReactElement;
 }