+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage/theme
 
+## 0.6.7-next.1
+
+### Patch Changes
+
+- 373486e: Add a different background colour to focused `MenuItem`s to fix a bug in MUI 4 (which is fixed in MUI 5)
+
 ## 0.6.7-next.0
 
 ### Patch Changes